# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API Todo List",
        "description": "Dokumentasi api yang tersedia",
        "contact": {
            "email": "iqbal.ramadhani@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/show-all-checklist": {
            "get": {
                "tags": [
                    "Check List"
                ],
                "summary": "Show All Checklist",
                "description": "Show All Checklist",
                "operationId": "351bf36abfcb2a2c156a2ef0adb30ace",
                "responses": {
                    "401": {
                        "description": "Error validate access token or form data",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "string"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error Internal Server",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "error": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/add-checklist": {
            "post": {
                "tags": [
                    "Check List"
                ],
                "summary": "Add Checklist",
                "description": "",
                "operationId": "8ad05206c7bdac2238e3eddb41c382d6",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Judul",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error Data Validation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error Auth Validation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error Internal Server",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "error": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/remove-checklist": {
            "delete": {
                "tags": [
                    "Check List"
                ],
                "summary": "Remove Checklist",
                "description": "",
                "operationId": "102e6ed052f4b8819305d11fd1d927de",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "ID Checklist",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error Data Validation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error Auth Validation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error Internal Server",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "error": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}